requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
marshmallow-sqlalchemy==0.26.1
sqlalchemy-redshift==0.8.14
requests==2.31.0
pandas==1.4.2
psycopg2-binary==2.9.5